Add a check to ensure a sending status changes in a cycle.
